Apple support

+++ WickedEngine/wiScene_Components.cpp
@@ -853,17 +853,17 @@ namespace wi::scene
 		}
 		const uint64_t alignment = device->GetMinOffsetAlignment(&bd);
 		bd.size =
-			AlignTo(vertex_positions.size() * position_stride, alignment) + // position will be first to have 0 offset for flexible alignment!
-			AlignTo(provoke.size() * GetProvokingIndexStride(), alignment) +
-			AlignTo(reorder.size() * GetIndexStride(), alignment) +
-			AlignTo(indices.size() * GetIndexStride(), alignment) +
-			AlignTo(vertex_normals.size() * sizeof(Vertex_NOR), alignment) +
-			AlignTo(vertex_tangents.size() * sizeof(Vertex_TAN), alignment) +
-			AlignTo(uv_count * uv_stride, alignment) +
-			AlignTo(vertex_atlas.size() * sizeof(Vertex_TEX), alignment) +
-			AlignTo(vertex_colors.size() * sizeof(Vertex_COL), alignment) +
-			AlignTo(vertex_boneindices.size() * sizeof(Vertex_BON), alignment) +
-			AlignTo(vertex_boneindices2.size() * sizeof(Vertex_BON), alignment)
+			AlignTo(uint64_t(vertex_positions.size() * position_stride), alignment) + // position will be first to have 0 offset for flexible alignment!
+			AlignTo(uint64_t(provoke.size() * GetProvokingIndexStride()), alignment) +
+			AlignTo(uint64_t(reorder.size() * GetIndexStride()), alignment) +
+			AlignTo(uint64_t(indices.size() * GetIndexStride()), alignment) +
+			AlignTo(uint64_t(vertex_normals.size() * sizeof(Vertex_NOR)), alignment) +
+			AlignTo(uint64_t(vertex_tangents.size() * sizeof(Vertex_TAN)), alignment) +
+			AlignTo(uint64_t(uv_count * uv_stride), alignment) +
+			AlignTo(uint64_t(vertex_atlas.size() * sizeof(Vertex_TEX)), alignment) +
+			AlignTo(uint64_t(vertex_colors.size() * sizeof(Vertex_COL)), alignment) +
+			AlignTo(uint64_t(vertex_boneindices.size() * sizeof(Vertex_BON)), alignment) +
+			AlignTo(uint64_t(vertex_boneindices2.size() * sizeof(Vertex_BON)), alignment)
 			;
 
 		constexpr Format morph_format = Format::R16G16B16A16_FLOAT;
@@ -872,11 +872,11 @@ namespace wi::scene
 		{
 			if (!morph.vertex_positions.empty())
 			{
-				bd.size += AlignTo(vertex_positions.size() * morph_stride, alignment);
+				bd.size += AlignTo(uint64_t(vertex_positions.size() * morph_stride), alignment);
 			}
 			if (!morph.vertex_normals.empty())
 			{
-				bd.size += AlignTo(vertex_normals.size() * morph_stride, alignment);
+				bd.size += AlignTo(uint64_t(vertex_normals.size() * morph_stride), alignment);
 			}
 		}
 
@@ -983,11 +983,11 @@ namespace wi::scene
 				}
 			}
 
-			bd.size = AlignTo(bd.size, sizeof(ShaderCluster));
-			bd.size = AlignTo(bd.size + clusters.size() * sizeof(ShaderCluster), alignment);
+			bd.size = AlignTo(bd.size, uint64_t(sizeof(ShaderCluster)));
+			bd.size = AlignTo(uint64_t(bd.size + clusters.size() * sizeof(ShaderCluster)), alignment);
 
-			bd.size = AlignTo(bd.size, sizeof(ShaderClusterBounds));
-			bd.size = AlignTo(bd.size + cluster_bounds.size() * sizeof(ShaderClusterBounds), alignment);
+			bd.size = AlignTo(bd.size, uint64_t(sizeof(ShaderClusterBounds)));
+			bd.size = AlignTo(uint64_t(bd.size + cluster_bounds.size() * sizeof(ShaderClusterBounds)), alignment);
 		}
 
 		auto init_callback = [&](void* dest) {
@@ -1301,7 +1301,7 @@ namespace wi::scene
 								XMStoreHalf4(vertices + ind, XMLoadFloat3(&morph.vertex_positions[i]));
 							}
 						}
-						buffer_offset += AlignTo(morph.vertex_positions.size() * sizeof(XMHALF4), alignment);
+						buffer_offset += AlignTo(uint64_t(morph.vertex_positions.size() * sizeof(XMHALF4)), alignment);
 					}
 					if (!morph.vertex_normals.empty())
 					{
@@ -1325,7 +1325,7 @@ namespace wi::scene
 								XMStoreHalf4(vertices + ind, XMLoadFloat3(&morph.vertex_normals[i]));
 							}
 						}
-						buffer_offset += AlignTo(morph.vertex_normals.size() * sizeof(XMHALF4), alignment);
+						buffer_offset += AlignTo(uint64_t(morph.vertex_normals.size() * sizeof(XMHALF4)), alignment);
 					}
 				}
 				vb_mor.size = buffer_offset - vb_mor.offset;
@@ -1333,7 +1333,7 @@ namespace wi::scene
 
 			if (!clusters.empty())
 			{
-				buffer_offset = AlignTo(buffer_offset, sizeof(ShaderCluster));
+				buffer_offset = AlignTo(buffer_offset, uint64_t(sizeof(ShaderCluster)));
 				vb_clu.offset = buffer_offset;
 				vb_clu.size = clusters.size() * sizeof(ShaderCluster);
 				std::memcpy(buffer_data + buffer_offset, clusters.data(), vb_clu.size);
@@ -1341,7 +1341,7 @@ namespace wi::scene
 			}
 			if (!cluster_bounds.empty())
 			{
-				buffer_offset = AlignTo(buffer_offset, sizeof(ShaderClusterBounds));
+				buffer_offset = AlignTo(buffer_offset, uint64_t(sizeof(ShaderClusterBounds)));
 				vb_bou.offset = buffer_offset;
 				vb_bou.size = cluster_bounds.size() * sizeof(ShaderClusterBounds);
 				std::memcpy(buffer_data + buffer_offset, cluster_bounds.data(), vb_bou.size);
@@ -1451,10 +1451,10 @@ namespace wi::scene
 
 		const uint64_t alignment = device->GetMinOffsetAlignment(&desc) * sizeof(Vertex_POS32); // additional alignment for RGB32F
 		desc.size =
-			AlignTo(vertex_positions.size() * sizeof(Vertex_POS32W), alignment) + // pos
-			AlignTo(vertex_positions.size() * sizeof(Vertex_POS32W), alignment) + // prevpos
-			AlignTo(vertex_normals.size() * sizeof(Vertex_NOR), alignment) +
-			AlignTo(vertex_tangents.size() * sizeof(Vertex_TAN), alignment)
+			AlignTo(uint64_t(vertex_positions.size() * sizeof(Vertex_POS32W)), alignment) + // pos
+			AlignTo(uint64_t(vertex_positions.size() * sizeof(Vertex_POS32W)), alignment) + // prevpos
+			AlignTo(uint64_t(vertex_normals.size() * sizeof(Vertex_NOR)), alignment) +
+			AlignTo(uint64_t(vertex_tangents.size() * sizeof(Vertex_TAN)), alignment)
 			;
 
 		bool success = device->CreateBuffer(&desc, nullptr, &streamoutBuffer);
